+++ src/main/java/com/kazakhi/onlinebookshop/service/CartService.java
@@ -8,7 +8,7 @@ import java.util.List;
 @Service
 public interface CartService {
     List<Cart> getCartItems(Integer userId);
-    Cart addToCart(Long userId, Integer bookId, Integer quantity);
-    Cart updateCart(Integer userId, Integer bookId, Integer quantity);
-    void removeFromCart(Integer userId, Integer bookId);
+    Cart addToCart(Integer userId, Long bookId, Integer quantity);
+    Cart updateCart(Integer userId, Long bookId, Integer quantity);
+    void removeFromCart(Integer userId, Long bookId);
 }
